feat(Google Cloud Storage loader) — addressing feedback
